+++ project2/GIPHYAPI.py
@@ -1,20 +1,15 @@
-import click
+import os
 
+#this will access the api key
+GIPHY_API_KEY = os.environ.get("GIPHY_API_KEY")
 
-@click.group()
-def gif():
-    print("hello from giphy cli!")
-
-
-@gif.command()
-def trending():
-    print("trending subcommand called!")
-
-
-@gif.command()
-def search():
-    print("search subcommand called!")
-
-
-if __name__ == "__main__":
-    gif()
+if GIPHY_API_KEY is None:
+    print("error... I don't think u set the api key")
+else:
+    def get_trending(api_key):
+    url = "https://api.giphy.com/v1/gifs/trending?api_key={api_key}&limit=25&offset=0&rating=g"
+    response = requests.get(url)
+    # data = response.json(url)
+    data = response.json()
+    return data
+
